requirements.txt
playwright==1.51.0
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-xdist==3.6.1
python-dotenv==1.1.0
pytest==8.3.5
pytest-playwright==0.7.0
python-dotenv==1.1.0
pytest-html==4.1.1
allure-pytest==2.14.0
pytest-xdist==3.6.1
PyYAML==6.0.2
pytest-base-url==2.1.0
mysql-connector-python==9.3.0
ruff==0.11.7
black==25.1.0
isort==6.0.1
requests==2.32.3
pre_commit==4.2.0
pyodbc==5.2.0
